pyproject.toml
[project]
name = "ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.3.0",
    "langchain-google-genai>=0.3.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0.0",
    "pymupdf>=1.24.0",
]

[project.optional-dependencies]
openai = [
    "langchain-openai>=0.3.0",
]
anthropic = [
    "langchain-anthropic>=0.3.0",
]
dev = [
    "pytest",
    "ruff",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]